Refactoring CRN representation classes
Pretty big refactor to most of the classes that represent stuff in the CRN simulator.

Mainly done so that the simulator can hold a representation of the CRN itself, which it should need to recompute reaction probabilities if it updates the count of K between batches.

Still lots of placeholder values in various spots (not intended to be used in current form).