Implement continuous time
For now, we're assuming that the hypoexponential is always equal to its mean, and that its mean is well approximated by pretending that all of the rates are equal to their geometric mean.

There could easily be a lot of bugs, but lotka_volterra_testing and lotka_volterra_plotting are generally working.